apps/api/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
black==25.1.0
CacheControl==0.14.2
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
Deprecated==1.2.18
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.11
fastapi-cli==0.0.7
firebase-admin==6.6.0
google-api-core==2.24.1
google-api-python-client==2.162.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-appengine-admin==1.14.1
google-cloud-appengine-logging==1.6.1
google-cloud-audit-log==0.3.2
google-cloud-build==3.31.1
google-cloud-core==2.4.2
google-cloud-firestore==2.20.1
google-cloud-logging==3.11.4
google-cloud-monitoring==2.27.1
google-cloud-storage==3.1.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.69.0
greenlet==3.1.1
grpc-google-iam-v1==0.14.2
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
idna==3.10
importlib_metadata==8.6.1
Jinja2==3.1.5
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
motor==3.7.0
msgpack==1.1.0
mypy-extensions==1.0.0
opentelemetry-api==1.31.1
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
playwright==1.50.0
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pyee==12.1.1
Pygments==2.19.1
PyJWT==2.10.1
pymongo==4.11.1
pyparsing==3.2.1
python-dotenv==1.0.1
python-http-client==3.3.7
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
rich-toolkit==0.13.2
rsa==4.9
sendgrid==6.11.0
shellingham==1.5.4
sniffio==1.3.1
starkbank-ecdsa==2.2.0
starlette==0.46.0
tomli==2.2.1
typer==0.15.2
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==15.0
wrapt==1.17.2
zipp==3.21.0
